fix archive and unarchive bugs, add exception if existing tar file, improve errors messages text

+++ TheCrazyArchiver.py
@@ -7,6 +7,7 @@ import shutil
 from exception.ExceptionMissingFile import ExceptionMissingFile
 from exception.ExceptionNbLoop import ExceptionNbLoop
 from exception.ExceptionExistingFile import ExceptionExistingFile
+from exception.ExceptionExistingTar import ExceptionExistingTar
 
 
 class TheCrazyArchiver:
@@ -23,6 +24,9 @@ class TheCrazyArchiver:
             file_name = os.path.basename(file_path)
             file_path_dir = os.path.dirname(file_path)
             tar_name = file_name + ".tar"
+            tar_path = file_path_dir + "\\" + tar_name
+            if os.path.exists(tar_path):
+                raise ExceptionExistingTar(tar_path)
             count = 1
             while count <= nb_loop:
                 tar_path = file_path_dir + "\\" + tar_name + "_1"
@@ -43,12 +47,14 @@ class TheCrazyArchiver:
     def unarchive(file):
         file_path = os.path.realpath(file)
         tar = tarfile.open(file)
-        path_tmp = os.path.dirname(os.path.realpath(file)) + "\\directory\\"
+        path_tmp = os.path.dirname(os.path.realpath(file))
         tar.close()
         file_hide_path = TheCrazyArchiver.__unarchive(path_tmp)
+        path_tmp += "\\directory"
         file_hide_path_2 = os.path.dirname(file_path) + "\\" + os.path.basename(file_hide_path)
         if os.path.exists(file_hide_path_2):
-            shutil.rmtree(path_tmp)
+            if os.path.exists(path_tmp):
+                shutil.rmtree(path_tmp)
             raise ExceptionExistingFile(file_hide_path_2)
         os.rename(file_hide_path, file_hide_path_2)
         shutil.rmtree(path_tmp)
@@ -56,7 +62,7 @@ class TheCrazyArchiver:
     @staticmethod
     def __unarchive(path_dir_init):
         files_list = os.listdir(path_dir_init)
-        path_before = path_dir_init
+        path_before = path_dir_init + "\\"
         file_hide_path = ""
         for file in files_list:
             if os.path.isfile(path_before + file):
@@ -71,7 +77,8 @@ class TheCrazyArchiver:
                         if f.endswith('.tar'):
                             file_hide_path = TheCrazyArchiver.__unarchive(path_after)
                 except tarfile.ReadError:
-                    print("Fin de l'extraction du fichier : " + file)
-                    file_hide_path = os.path.realpath(path_before + file)
+                    if file_hide_path is "":
+                        print("Fin de l'extraction du fichier : " + file)
+                        file_hide_path = os.path.realpath(path_before + file)
                     return file_hide_path
         return file_hide_path
